feat(capture): harden screen-region source identity
