+++ src/backend/wayland/state/data.rs
@@ -5,6 +5,7 @@ use crate::backend::wayland::toolbar::hit::HitRegion;
 use crate::backend::wayland::zoom::ZoomWaiterRegistry;
 
 use super::region_capture::ActiveScreenRegion;
+use super::screen_image::ScreenSourceToken;
 
 use super::capture::OverlayCaptureBarrier;
 
@@ -142,6 +143,7 @@ pub struct StateData {
     pub(super) pending_freeze_on_start: bool,
     pub(super) screen_acquisition: ScreenAcquisitionRegistry,
     pub(super) zoom_waiter: ZoomWaiterRegistry,
+    pub(super) active_eyedropper_source: Option<ScreenSourceToken>,
     pub(super) active_screen_region: Option<ActiveScreenRegion>,
     pub(super) next_screen_region_generation: u64,
     pub(super) frozen_enabled: bool,
@@ -249,6 +251,7 @@ impl StateData {
             pending_freeze_on_start: false,
             screen_acquisition: ScreenAcquisitionRegistry::default(),
             zoom_waiter: ZoomWaiterRegistry::default(),
+            active_eyedropper_source: None,
             active_screen_region: None,
             next_screen_region_generation: 1,
             frozen_enabled: false,
